// app/types.ts
export interface Board {
  author: string;
  content: string;
  hit: number;
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface Notice {
  author: string;
  content: string;
  hit: number;
  id: number;
  title: string;
  updatedAt: string;
  attached: string;
}

export interface BoardInput {
  category: string;
  title: string;
  thumbnail: string;
  content: string;
  attached: string;
}